Obsolete static unknown element factory
We need access to namespace information, which is most reliably retrieved from a part container.
